pyproject.toml
[project]
name = "noctua"
version = "0.1.0"
description = "AI-powered RSS feed aggregator with intelligent summarization"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Your Name", email = "your@email.com" }]
keywords = ["rss", "feed", "aggregator", "ai", "summarization"]

dependencies = [
    "feedparser>=6.0.0",      # RSS/Atom feed parsing
    "httpx>=0.27.0",          # Async HTTP client
    "pyyaml>=6.0",            # YAML config parsing
    "beautifulsoup4>=4.12.0", # HTML parsing and cleanup
    "lxml>=5.0.0",            # Fast XML/HTML parser
    "python-dateutil>=2.9.0", # Date parsing utilities
    "pydantic>=2.0.0",        # Data validation
    "rich>=13.0.0",           # Beautiful terminal output
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
